requirement.txt
django==3.0.4
djangorestframework==3.12.2